### 2021-05-05 #113 @whedon
[PRE REVIEW]: OpenQuestion: A survey building and reporting platform written in Python
**Submitting author:** @Alcampopiano (<a href="http://orcid.org/0000-0002-3280-4447">Allan Campopiano</a>)
**Repository:** <a href="https://github.com/Alcampopiano/OpenQuestion" target ="_blank">https://github.com/Alcampopiano/OpenQuestion</a>
**Version:** v0.1.0

@Alcampopiano if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/jose-reviewers) have already agreed to review for JOSE and may be suitable for this submission.
