### 2021-07-25 #113 @labarba
hi @stefania11 👋 — You signed up to review for JOSE (The Journal of Open Source Education) a while back. Do you think you could help us with this submission?
**OpenQuestion: A survey building and reporting platform written in Python**
Target repository: https://github.com/Alcampopiano/OpenQuestion
Paper: https://github.com/openjournals/jose-papers/blob/jose.00113/jose.00113/10.21105.jose.00113.pdf

It’s a Python tool to build surveys and issue reports. It uses Anvil, Python, and JSON.

Let me know if you're interested in contributing a review for this. Thank you!
