### 2021-07-27 #113 @stefania11
Sounds good!

On Mon, Jul 26, 2021, 4:57 PM Lorena A. Barba ***@***.***>
wrote:

> Thank you, Stefania! After I add a second reviewer we can start the review
> issue, which will include a checklist to guide the review. In the meantime,
> you can of course look at the contents of the submission, and ask me any
> questions you may have.
>
