### 2021-07-28 #123 @whedon
[REVIEW]: OpenQuestion: A survey building and reporting platform written in Python
**Submitting author:** @Alcampopiano (<a href="http://orcid.org/0000-0002-3280-4447">Allan Campopiano</a>)
**Repository:** <a href="https://github.com/Alcampopiano/OpenQuestion" target ="_blank">https://github.com/Alcampopiano/OpenQuestion</a>
**Version:** v0.1.0

Due to the challenges of the COVID-19 pandemic, JOES is currently operating in a "reduced service mode".

- [x ] **License:** Does the repository contain a plain-text LICENSE file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license?
- [ ] **Version:** Does the release version given match the GitHub release (v0.1.0)? - "No releases published"

-  The installation instructions are minimal, insufficient, and hard to find ( it seems that the project requires the anvil-app-server package to test installation. Authors link some external documentation for how to install this package but the links are confusing and one cannot test the functionality of this repo without spending extensive time installing the anvil server infrastructure
- [ ] no due to lack of clear installation instructions

- [ ] **Installation instructions:** Is there a clearly stated list of dependencies? (Ideally, these should be handled with an automated package management solution.)
