### 2021-07-28 #123 @labarba
👋  @stefania11, @csev —  Here is the Review issue for "OpenQuestion: A survey building and reporting platform written in Python." The article proof is linked on the comment just above, and the submission repository is linked at the very top of this thread. You both have a reviewer checklist to guide you, review criteria are listed in our [docs](https://openjournals.readthedocs.io/en/jose/review_criteria.html) and you an ask me anything here!

You should feel free to open an issue in the target repository if needed, adding a link to this review issue to get the cross-link show up here.
