### 2021-11-03 #123 @csev
@Alcampopiano - I am catching up with this - I tried to install this with pip on a Mac and got "ERROR: Failed building wheel for setproctitle" - if there is a workaround that you know of, let me know.  I can spin up a docker and try again in a fresh linux but that will take some time.  I just want to run through the install process and get it to come up.
