### 2022-08-12 #123 @labarba
Hi all,
I have communicated with `@stefania11` via email regarding this review, and she has stated that, after her first review of this work some months ago, she has no more bandwidth for it now. I also note that she entered review comments by editing the checklist post at the top of this issue thread, noting the following under **Installation**:

> The installation instructions are minimal, insufficient, and hard to find ( it seems that the project requires the anvil-app-server package to test installation. Authors link some external documentation for how to install this package but the links are confusing and one cannot test the functionality of this repo without spending extensive time installing the anvil server infrastructure.

I also notice that @csev entered the following comment above:

> I tried to install this with pip on a Mac and got "ERROR: Failed building wheel for setproctitle" - if there is a workaround that you know of, let me know. I can spin up a docker and try again in a fresh linux but that will take some time. I just want to run through the install process and get it to come up.

Both reviewers therefor point to serious barriers for reuse of this software, while we did not have much engagement from the author. Therefore, I am making an editorial decision to reject this submission for publication in JOSE. I appreciate the author's effort, intention, and work, but we look to publish on software or lessons that are ready for reuse by others.
