### 2022-08-12 #123 @Alcampopiano
Hi @labarba 👋,

Thank you for jumping in here. I would please like to request a bit of grace.

I have been following this submission very closely. Over covid things really slowed down on JOSE (and everywhere) and I was patiently waiting for the reviewers to complete their initial review. They were pinged multiple times without response (totally understandable) and so I was happy to wait for the initial review to be completed.

I honestly did not realize there was action required of me as the reviewers were asked to come back to the thread to finish up (but I should have been explicitly present and for that I apologize).

> hiya, @stefania11, @csev – could you pop in this review thread again? We’re hoping this can move now, as it’s been quite some time. Thank you! :pray:

Regarding the installation issues. I believe these were related to the M1 Mac and have now been resolved. To demonstrate this, I've recorded a Loom video below of me installing the software by following the installation instructions in the documentation.

https://www.loom.com/share/ccea5aff0c97413a9beedd58ed7424a8

Since I can demonstrate that the installation does work as indicated in the documentation, may I have a bit of grace on the communication issues? As I mentioned, I misunderstood that things were waiting exclusively on me.

This project and review has my full attention but I misunderstood that a reject/accept was hanging in the balance, waiting on me. I think that is a reasonable assumption on my part but I should have been more explicit with you all.

Please let me know what you think.
