### 2022-08-23 #123 @csev
@Alcampopiano - I am happy to pick this back up - especially if the install process works.   I really don't like to say "yes" to software submissions that breaks immediately :) But if we can work through this and I can verify it installs from the repo and readme - I can be thumbs up.  Having a server to test is nice but not sufficient for me if the install from documentation / README seems to be broken.
