### 2022-08-23 #123 @Alcampopiano
@csev I really appreciate your message and agree completely. I'm confident that installation can be a smooth process.

In part, this is because OpenQuestion is built on top of [Anvil's open-source runtime engine](https://anvil.works/open-source) which powers thousands of real-world applications (open source and commercial). If we experience any issues, I'm certain that we can resolve them since we would not be the only ones affected.

Having said that, I must do my part to explain the instructions **very** clearly. I've recorded [a short video](https://www.loom.com/share/ccea5aff0c97413a9beedd58ed7424a8) above where I go through the installation [from the docs](https://alcampopiano.github.io/OpenQuestion/installation/) line by line.

@csev If you are willing, could you try the following instructions again (with a clean environment)? Please let me know if I can help in any way. I will be very present.

For convenience, here are the steps from the docs:

```
# install from pip
pip install OpenQuestion

# clone the OpenQuestion project into your current working directory
git clone https://github.com/Alcampopiano/OpenQuestion.git

# open up a python shell to interact with app's backend
anvil-app-server --app OpenQuestion --shell

# add a new row to the Users table
new_user=app_tables.users.add_row(email='your_username@example.com', enabled=True)

# add an encrypted password
# these steps are not needed if using Google or Microsoft Authentication
import bcrypt
password_hash = bcrypt.hashpw(b'new password', bcrypt.gensalt(16))
new_user['password_hash'] = password_hash.decode()

# leave the Python shell (this is not in the docs but I will add it)
quit()

# launch the app itself!
anvil-app-server --app OpenQuestion
```

Sometimes there are eccentricities with local operating systems but on my mac M1 and Linux machine, these steps seem to work as expected. I will work to resolve any issues that come up.

Thank you again for your response Charles. I really appreciate it.
