### 2022-08-23 #123 @labarba
@Alcampopiano – Since I rejected this submission on the basis of the broken installation, you would have to make a new submission once you iron things out. As I noted [above](https://github.com/openjournals/jose-reviews/issues/123#issuecomment-12130194410, I communicated with reviewer `@stefania11` via email, and after the unhappy installation attempt, this reviewer is no longer willing / available to review. And that is what happens too with potential users if they run into hard installation troubles! The goal of a JOSE submission of for the software or teaching module to be reusable by others.
