### 2022-08-23 #123 @Alcampopiano
@labarba Thanks for responding. I strongly believe that the review that prompted the rejection was rather **unfair (and unclear)**.  I want to kindly bring this to your attention as I think it deserves addressed.

- In your opinion, do you think installation is too difficult (given [this short video](https://www.loom.com/share/ccea5aff0c97413a9beedd58ed7424a8) or [the steps listed below](https://github.com/openjournals/jose-reviews/issues/123#issuecomment-1224250795) from the docs)?

- May I please get your thoughts on what I believe was a rather unfair and unclear review? For example:

>  _installation is hard to find_

However, it's listed under "installation" in the docs ToC. This is not hard to find in my opinion. Where else should it be listed if not there?

> _Authors link some external documentation for how to install this package but the links are confusing_

This is not true or clearly stated. The installation instructions are completely self contained. A `pip` and a `git clone` are sufficient for installing everything needed and this is clearly stated in the docs.

> _...without spending extensive time installing the anvil server infrastructure_

This dependency is installed automatically with `pip` in the first line of the installation instructions. I do not understand what the reviewer means by "extensive time installing". There's no rule against having a dependency of course 🙂.

In addition: (1) reviewer 1's comments were rather hidden above (not posted as an issue and I wasn't tagged). However, I take some responsibility for not noticing this. (2) The second reviewer is happy to review and this counts for something I think 🙂.

@labarba I really appreciate all the work that goes into this reviewing and I apologize for the misunderstandings on my part during the process.

I kindly ask for your opinion on the review I received, especially given my response and demonstrations.
